Move main WeakMap logic to abstract class
